Nest McpError as Mcp.Error, dropping a dead duplicate payload
`object Mcp` already had a nested `Error` — a JSON-RPC error payload — which is
what blocked `McpError → Mcp.Error` in the second pass. That payload is dead: it
is declared once and referenced nowhere, and its three fields are exactly
`JsonRpc.Failure(code, message, data)`, which the same file already calls at the
one place it builds an error response.

So it is deleted rather than renamed, and `McpError` takes the name.
