+++ lib/synesthesia/src/core/soundness_synesthesia_core.scala
@@ -34,4 +34,4 @@ package soundness
 
 export
   synesthesia
-  . { about, agent, Agent, Discourse, human, Human, Mcp, McpError, prompt, resource, title, tool, ui }
+  . { about, agent, Agent, Discourse, human, Human, Mcp, prompt, resource, title, tool, ui }
